+++ TSClientGen.Core/transport-axios.ts
@@ -1,32 +1,29 @@
 import { RequestOptions, GetUriOptions } from './transport-contracts';
-import axios, { AxiosRequestConfig } from 'axios';
+import axios from 'axios';
 
-axios.defaults.headers.post['Content-Type'] = 'application/json';
-axios.defaults.headers.put['Content-Type'] = 'application/json';
+function getCancelToken(request: RequestOptions) {
+	return typeof request.getAbortFunc === 'function' ? new axios.CancelToken(request.getAbortFunc) : undefined;
+}
 
 export async function request<TResponse>(request: RequestOptions): Promise<TResponse> {
-	const options: AxiosRequestConfig = {
+	const response = await axios.request<TResponse>({
+		baseURL: request.baseURL,
 		url: request.url,
 		method: request.method,
+		headers: request.headers,
 		params: request.queryStringParams,
 		data: request.requestBody,
 		onUploadProgress: request.onUploadProgress,
+		cancelToken: getCancelToken(request),
 		timeout: request.timeout,
-		headers: request.headers,
-		baseURL: request.baseURL,
-	};
-	if (typeof request.getAbortFunc == 'function') {
-		options.cancelToken = new axios.CancelToken(request.getAbortFunc);
-	}
-	const response = await axios.request<TResponse>(options);
+	});
 	return response.data;
 }
 
-export const getUri: (options: GetUriOptions) => string = function (options: GetUriOptions) {
-	const axiosOptions: AxiosRequestConfig = {
+export function getUri(options: GetUriOptions): string {
+	return axios.getUri({
+		baseURL: options.baseURL,
 		url: options.url,
 		params: options.queryStringParams,
-	};
-
-	return axios.getUri(axiosOptions);
-};
+	});
+}